Major UI Fixes
Minimap is now working well, only in world 2 - 1 and 2 - 2 as it is in a "surface area".

ShopMenu:
- Animation is now smoother and faster
- Added Outro pullout anim.
- Potions can be interacted
- Added extra weapons just for the looks
- SFX for putting potion in cart (PotionBubbleSFX)

Remade Quest UI. Made it look prettier with some fancy boxes and animated sprites for the checkbox. No more Fragment text UI on the bottom left.
Repositioned the UI so that it alligns with minimap on the topleft and interactive DS4 gui on bottomright.